+++ src/domains/store/api/storeApi.js
@@ -1,13 +1,12 @@
-import axios from "axios"
+import apiInstance from "@/common/api/apiInstance.js";
 
-// TODO: URL을 입력해주세요
-const BASE_URL = "/api" // test용 instance + env로 변경하기
+
 
 //"http://localhost:8072/api/store"
 export async function createStore(data) {
-  const response = await axios.post(`${BASE_URL}/store`, data,{
+  const response = await apiInstance.post(`/store`, data,{
     headers: {
-      "X-User-Id": 504, // headers 지우기
+      "X-User-Id": 505, // headers 지우기
     }
   })
   return response.data
